Initial commit
This repo is just a quick test I did for generation of snapcraft.yaml files. I mostly am using it to generate version numbers on the fly (but maybe some other things if needs be)
to be run by jenkins and then continue to build the package directly. It pretty much builds a thing that prints hello right now and probably won't be updated afterwards. I'm just
making this to store for later when I needed it.